Refactor files.py into separate classes
