+++ src/zos_files/zowe/zos_files_for_zowe_sdk/__init__.py
@@ -4,3 +4,6 @@ Zowe Python SDK - z/OS Files package
 
 from . import constants, exceptions
 from .files import Files
+from .datasets import Datasets
+from .uss import USSFiles
+from .file_system import FileSystems
