+++ src/zos_files/zowe/zos_files_for_zowe_sdk/datasets.py
@@ -1,0 +1,614 @@
+"""Zowe Python Client SDK.
+
+This program and the accompanying materials are made available under the terms of the
+Eclipse Public License v2.0 which accompanies this distribution, and is available at
+
+https://www.eclipse.org/legal/epl-v20.html
+
+SPDX-License-Identifier: EPL-2.0
+
+Copyright Contributors to the Zowe Project.
+"""
+
+import os
+
+from zowe.core_for_zowe_sdk import SdkApi
+from zowe.core_for_zowe_sdk.exceptions import FileNotFound
+from zowe.zos_files_for_zowe_sdk import constants, exceptions
+from zowe.zos_files_for_zowe_sdk.constants import FileType, zos_file_constants
+
+_ZOWE_FILES_DEFAULT_ENCODING = "utf-8"
+
+class Datasets(SdkApi):
+    """
+    Class used to represent the base z/OSMF Datasets API
+    which includes all operations related to datasets.
+
+    ...
+
+    Attributes
+    ----------
+    connection
+        connection object
+    """
+
+    def __init__(self, connection):
+        """
+        Construct a Datasets object.
+
+        Parameters
+        ----------
+        connection
+            The z/OSMF connection object (generated by the ZoweSDK object)
+
+        Also update header to accept gzip encoded responses
+        """
+        super().__init__(connection, "/zosmf/restfiles/", logger_name=__name__)
+        self.default_headers["Accept-Encoding"] = "gzip"
+
+    def list(self, name_pattern, return_attributes=False):
+        """Retrieve a list of datasets based on a given pattern.
+
+        Parameters
+        ----------
+        name_pattern : str
+            The pattern to match dataset names.
+        return_attributes : bool, optional
+            Whether to return dataset attributes along with the names. Defaults to False.
+
+        Returns
+        -------
+            list of dict
+
+            A JSON with a list of dataset names (and attributes if specified) matching the given pattern.
+        """
+        custom_args = self._create_custom_request_arguments()
+        custom_args["params"] = {"dslevel": self._encode_uri_component(name_pattern)}
+        custom_args["url"] = "{}ds".format(self.request_endpoint)
+
+        if return_attributes:
+            custom_args["headers"]["X-IBM-Attributes"] = "base"
+
+        response_json = self.request_handler.perform_request("GET", custom_args)
+        return response_json
+    
+    def list_members(self, dataset_name, member_pattern=None, member_start=None, limit=1000, attributes="member"):
+        """Retrieve the list of members on a given PDS/PDSE.
+
+        Returns
+        -------
+        json
+            A JSON with a list of members from a given PDS/PDSE
+        """
+        custom_args = self._create_custom_request_arguments()
+        additional_parms = {}
+        if member_start is not None:
+            additional_parms["start"] = member_start
+        if member_pattern is not None:
+            additional_parms["pattern"] = member_pattern
+        custom_args["params"] = additional_parms
+        custom_args["url"] = "{}ds/{}/member".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        custom_args["headers"]["X-IBM-Max-Items"] = "{}".format(limit)
+        custom_args["headers"]["X-IBM-Attributes"] = attributes
+        response_json = self.request_handler.perform_request("GET", custom_args)
+        return response_json["items"]  # type: ignore
+    
+    def copy_dataset_or_member(
+        self,
+        from_dataset_name,
+        to_dataset_name,
+        from_member_name=None,
+        volser=None,
+        alias=None,
+        to_member_name=None,
+        enq=None,
+        replace=False,
+    ):
+        """
+        Copy a dataset or member to another dataset or member.
+        Parameters
+        ----------
+        from_dataset_name: str
+            Name of the dataset to copy from
+        to_dataset_name: str
+            Name of the dataset to copy to
+        from_member_name: str
+            Name of the member to copy from
+        volser: str
+            Volume serial number of the dataset to copy from
+        alias: bool
+            Alias of the dataset to copy from
+        to_member_name: str
+            Name of the member to copy to
+        enq: str
+            Enqueue type for the dataset to copy from
+        replace: bool
+            If true, members in the target data set are replaced.
+        Returns
+        -------
+        json
+            A JSON containing the result of the operation
+        """
+
+        data = {
+            "request": "copy",
+            "from-dataset": {"dsn": from_dataset_name.strip(), "member": from_member_name},
+            "replace": replace,
+        }
+
+        path_to_member = f"{to_dataset_name}({to_member_name})" if to_member_name else to_dataset_name
+        if enq:
+            if enq in ("SHR", "SHRW", "EXCLU"):
+                data["enq"] = enq
+            else:
+                self.logger.error("Invalid value for enq.")
+                raise ValueError("Invalid value for enq.")
+        if volser:
+            data["from-dataset"]["volser"] = volser
+        if alias is not None:  # because it can be false so
+            data["from-dataset"]["alias"] = alias
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(path_to_member))
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+    
+    def get_content(self, dataset_name, stream=False):
+        """Retrieve the contents of a given dataset.
+
+        Returns
+        -------
+        json
+            A JSON with the contents of a given dataset
+        """
+        custom_args = self._create_custom_request_arguments()
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        response_json = self.request_handler.perform_request("GET", custom_args, stream=stream)
+        return response_json
+    
+    def create(self, dataset_name, options={}):
+        """
+        Create a sequential or partitioned dataset.
+        Parameters
+        ----------
+            dataset_name
+        Returns
+        -------
+        json
+        """
+
+        if options.get("like") is None:
+            if options.get("primary") is None or options.get("lrecl") is None:
+                self.logger.error("If 'like' is not specified, you must specify 'primary' or 'lrecl'.")
+                raise ValueError("If 'like' is not specified, you must specify 'primary' or 'lrecl'.")
+
+            for opt in (
+                "volser",
+                "unit",
+                "dsorg",
+                "alcunit",
+                "primary",
+                "secondary",
+                "dirblk",
+                "avgblk",
+                "recfm",
+                "blksize",
+                "lrecl",
+                "storclass",
+                "mgntclass",
+                "dataclass",
+                "dsntype",
+                "like",
+            ):
+                if opt == "dsorg":
+                    if options.get(opt) is not None and options[opt] not in ("PO", "PS"):
+                        self.logger.error(f"{opt} is not 'PO' or 'PS'.")
+                        raise KeyError
+
+                elif opt == "alcunit":
+                    if options.get(opt) is None:
+                        options[opt] = "TRK"
+                    else:
+                        if options[opt] not in ("CYL", "TRK"):
+                            self.logger.error(f"{opt} is not 'CYL' or 'TRK'.")
+                            raise KeyError
+
+                elif opt == "primary":
+                    if options.get(opt) is not None:
+                        if options["primary"] > 16777215:
+                            self.logger.error("Specified value exceeds limit.")
+                            raise ValueError
+
+                elif opt == "secondary":
+                    if options.get("primary") is not None:
+                        if options.get(opt) is None:
+                            options["secondary"] = int(options["primary"] / 10)
+                        if options["secondary"] > 16777215:
+                            self.logger.error("Specified value exceeds limit.")
+                            raise ValueError
+
+                elif opt == "dirblk":
+                    if options.get(opt) is not None:
+                        if options.get("dsorg") == "PS":
+                            if options["dirblk"] != 0:
+                                self.logger.error("Can't allocate directory blocks for files.")
+                                raise ValueError
+                        elif options.get("dsorg") == "PO":
+                            if options["dirblk"] == 0:
+                                self.logger.error("Can't allocate empty directory blocks.")
+                                raise ValueError
+
+                elif opt == "recfm":
+                    if options.get(opt) is None:
+                        options[opt] = "F"
+                    else:
+                        if options[opt] not in ("F", "FB", "V", "VB", "U", "FBA", "FBM", "VBA", "VBM"):
+                            self.logger.error("Invalid record format.")
+                            raise KeyError
+
+                elif opt == "blksize":
+                    if options.get(opt) is None and options.get("lrecl") is not None:
+                        options[opt] = options["lrecl"]
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        custom_args["json"] = options
+        response_json = self.request_handler.perform_request("POST", custom_args, expected_code=[201])
+        return response_json
+
+    def create_default(self, dataset_name: str, default_type: str):
+        """
+        Create a dataset with default options set.
+        Default options depend on the requested type.
+
+        Parameters
+        ----------
+            dataset_name: str
+            default_type: str
+                "partitioned", "sequential", "classic", "c" or "binary"
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        if default_type not in ("partitioned", "sequential", "classic", "c", "binary"):
+            self.logger.error("Invalid type for default data set.")
+            raise ValueError("Invalid type for default data set.")
+
+        custom_args = self._create_custom_request_arguments()
+
+        if default_type == "partitioned":
+            custom_args["json"] = {
+                "alcunit": "CYL",
+                "dsorg": "PO",
+                "primary": 1,
+                "dirblk": 5,
+                "recfm": "FB",
+                "blksize": 6160,
+                "lrecl": 80,
+            }
+        elif default_type == "sequential":
+            custom_args["json"] = {
+                "alcunit": "CYL",
+                "dsorg": "PS",
+                "primary": 1,
+                "recfm": "FB",
+                "blksize": 6160,
+                "lrecl": 80,
+            }
+        elif default_type == "classic":
+            custom_args["json"] = {
+                "alcunit": "CYL",
+                "dsorg": "PO",
+                "primary": 1,
+                "recfm": "FB",
+                "blksize": 6160,
+                "lrecl": 80,
+                "dirblk": 25,
+            }
+        elif default_type == "c":
+            custom_args["json"] = {
+                "dsorg": "PO",
+                "alcunit": "CYL",
+                "primary": 1,
+                "recfm": "VB",
+                "blksize": 32760,
+                "lrecl": 260,
+                "dirblk": 25,
+            }
+        elif default_type == "binary":
+            custom_args["json"] = {
+                "dsorg": "PO",
+                "alcunit": "CYL",
+                "primary": 10,
+                "recfm": "U",
+                "blksize": 27998,
+                "lrecl": 27998,
+                "dirblk": 25,
+            }
+
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        response_json = self.request_handler.perform_request("POST", custom_args, expected_code=[201])
+        return response_json
+
+    def get_binary_content(self, dataset_name, stream = False, with_prefixes=False):
+        """
+        Retrieve the contents of a given dataset as a binary bytes object.
+
+        Parameters
+        ----------
+        dataset_name: str - Name of the dataset to retrieve
+        with_prefixes: boolean - if True include a 4 byte big endian record len prefix
+                                 default: False
+        Returns
+        -------
+        response
+            A response object from the requests library
+        """
+        custom_args = self._create_custom_request_arguments()
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        custom_args["headers"]["Accept"] = "application/octet-stream"
+        if with_prefixes:
+            custom_args["headers"]["X-IBM-Data-Type"] = "record"
+        else:
+            custom_args["headers"]["X-IBM-Data-Type"] = "binary"
+        response = self.request_handler.perform_request("GET", custom_args, stream=stream)
+        return response
+
+    def write(self, dataset_name, data, encoding=_ZOWE_FILES_DEFAULT_ENCODING):
+        """Write content to an existing dataset.
+
+        Returns
+        -------
+        json
+            A JSON containing the result of the operation
+        """
+        custom_args = self._create_custom_request_arguments()
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        custom_args["data"] = data
+        custom_args["headers"]["Content-Type"] = "text/plain; charset={}".format(encoding)
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[204, 201])
+        return response_json
+
+    def download(self, dataset_name, output_file):
+        """Retrieve the contents of a dataset and saves it to a given file."""
+        response = self.get_content(dataset_name, stream = True)
+        with open(output_file, "w", encoding="utf-8") as f:
+            for chunk in response.iter_content(chunk_size=4096, decode_unicode=True):
+                f.write(chunk)
+
+    def download_binary(self, dataset_name, output_file, with_prefixes=False):
+        """Retrieve the contents of a binary dataset and saves it to a given file.
+
+        Parameters
+        ----------
+        dataset_name:str - Name of the dataset to download
+        output_file:str - Name of the local file to create
+        with_prefixes:boolean - If true, include a four big endian bytes record length prefix.
+                                The default is False
+        """
+        response = self.get_binary_content(dataset_name, with_prefixes=with_prefixes, stream=True)
+        with open(output_file, "wb") as f:
+            for chunk in response.iter_content(chunk_size=4096):
+                f.write(chunk)
+
+    def upload_file(self, input_file, dataset_name, encoding=_ZOWE_FILES_DEFAULT_ENCODING):
+        """Upload contents of a given file and uploads it to a dataset."""
+        if os.path.isfile(input_file):
+            with open(input_file, "rb") as in_file:
+                response_json = self.write(dataset_name, in_file)
+        else:
+            self.logger.error(f"File {input_file} not found.")
+            raise FileNotFound(input_file)
+        
+    def recall_migrated(self, dataset_name: str, wait=False):
+        """
+        Recalls a migrated data set.
+
+        Parameters
+        ----------
+        dataset_name: str
+            Name of the data set
+
+        wait: bool
+            If true, the function waits for completion of the request, otherwise the request is queued
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        data = {"request": "hrecall", "wait": wait}
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+
+    def delete_migrated(self, dataset_name: str, purge=False, wait=False):
+        """
+        Deletes migrated data set.
+
+        Parameters
+        ----------
+        dataset_name: str
+            Name of the data set
+
+        purge: bool
+            If true, the function uses the PURGE=YES on ARCHDEL request, otherwise it uses the PURGE=NO.
+
+        wait: bool
+            If true, the function waits for completion of the request, otherwise the request is queued.
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        data = {
+            "request": "hdelete",
+            "purge": purge,
+            "wait": wait,
+        }
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+
+    def migrate(self, dataset_name: str, wait=False):
+        """
+        Migrates the data set.
+
+        Parameters
+        ----------
+        dataset_name: str
+            Name of the data set
+
+        wait: bool
+            If true, the function waits for completion of the request, otherwise the request is queued.
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        data = {"request": "hmigrate", "wait": wait}
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+
+    def rename(self, before_dataset_name: str, after_dataset_name: str):
+        """
+        Renames the data set.
+
+        Parameters
+        ----------
+        before_dataset_name: str
+            The source data set name.
+
+        after_dataset_name: str
+            New name for the source data set.
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        data = {"request": "rename", "from-dataset": {"dsn": before_dataset_name.strip()}}
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(
+            self.request_endpoint, self._encode_uri_component(after_dataset_name).strip()
+        )
+
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+
+    def rename_member(self, dataset_name: str, before_member_name: str, after_member_name: str, enq=""):
+        """
+        Renames the data set member.
+
+        Parameters
+        ----------
+        dataset_name: str
+            Name of the data set.
+
+        before_member_name: str
+            The source member name.
+
+        after_member_name: str
+            New name for the source member.
+
+        enq: str
+            Values can be SHRW or EXCLU. SHRW is the default for PDS members, EXCLU otherwise.
+
+        Returns
+        -------
+        json - A JSON containing the result of the operation
+        """
+
+        data = {
+            "request": "rename",
+            "from-dataset": {
+                "dsn": dataset_name.strip(),
+                "member": before_member_name.strip(),
+            },
+        }
+
+        path_to_member = dataset_name.strip() + "(" + after_member_name.strip() + ")"
+
+        if enq:
+            if enq in ("SHRW", "EXCLU"):
+                data["enq"] = enq.strip()
+            else:
+                self.logger.error("Invalid value for enq.")
+                raise ValueError("Invalid value for enq.")
+
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(path_to_member))
+
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
+    
+    def delete(self, dataset_name, volume=None, member_name=None):
+        """Deletes a sequential or partitioned data."""
+        custom_args = self._create_custom_request_arguments()
+        if member_name is not None:
+            dataset_name = f"{dataset_name}({member_name})"
+        url = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(dataset_name))
+        if volume is not None:
+            url = "{}ds/-{}/{}".format(self.request_endpoint, volume, self._encode_uri_component(dataset_name))
+        custom_args["url"] = url
+        response_json = self.request_handler.perform_request("DELETE", custom_args, expected_code=[200, 202, 204])
+        return response_json
+    
+    def copy_uss_to_dataset(
+        self, from_filename, to_dataset_name, to_member_name=None, type=FileType.TEXT, replace=False
+    ):
+        """
+        Copy a USS file to dataset.
+
+        Parameters
+        ----------
+        from_filename: str
+            Name of the file to copy from.
+        to_dataset_name: str
+            Name of the dataset to copy to.
+        to_member_name: str
+            Name of the member to copy to.
+        type: FileType, optional
+            Type of the file to copy from. Default is FileType.TEXT.
+        replace: bool, optional
+            If true, members in the target dataset are replaced.
+
+        Returns
+        -------
+        json
+            A JSON containing the result of the operation.
+        """
+
+        data = {
+            "request": "copy",
+            "from-file": {"filename": from_filename.strip(), "type": type.value},
+            "replace": replace,
+        }
+
+        path_to_member = f"{to_dataset_name}({to_member_name})" if to_member_name else to_dataset_name
+        custom_args = self._create_custom_request_arguments()
+        custom_args["json"] = data
+        custom_args["url"] = "{}ds/{}".format(self.request_endpoint, self._encode_uri_component(path_to_member))
+        response_json = self.request_handler.perform_request("PUT", custom_args, expected_code=[200])
+        return response_json
